requirements.txt
streamlit==1.41.1
streamlit-authenticator==0.4.1
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
httpx==0.28.1
pyyaml==6.0.2
python-dateutil==2.9.0
bcrypt==4.2.1
apscheduler==3.10.4
lxml==5.3.0
pandas==2.2.3
